pyproject.toml
[project]
name = "abt-social-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Matt Holubinka",email = "matthew.holubinka@accenture.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "streamlit (>=1.56.0,<2.0.0)",
]

[tool.black]
target-version = ["py311"]
line-length = 88

[tool.ruff]

[tool.mypy]
python_version = "3.11"
exclude = ["tests"]
disallow_untyped_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = false
check_untyped_defs = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.10,<0.16.0)",
    "mypy (>=1.20.1,<2.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "httpx (>=0.27.0)",
    "types-requests (>=2.33.0.20260408,<3.0.0.0)",
]
